First compilable version with Minecraft 1.20.1

+++ src/main/java/baritone/process/FarmProcess.java
@@ -300,7 +300,7 @@ public final class FarmProcess extends BaritoneProcessHelper implements IFarmPro
         for (ItemEntity item : ctx.world().getEntitiesByClass(ItemEntity.class, ctx.entity().getBoundingBox().expand(30), Entity::isOnGround)) {
             if (PICKUP_DROPPED.contains(item.getStack().getItem())) {
                 // +0.1 because of farmland's 0.9375 dummy height lol
-                goalz.add(new GoalBlock(new BlockPos(item.getX(), item.getY() + 0.1, item.getZ())));
+                goalz.add(new GoalBlock(new BlockPos((int)item.getX(), (int)(item.getY() + 0.1), (int)(item.getZ()))));
             }
         }
         return new PathingCommand(new GoalComposite(goalz.toArray(new Goal[0])), PathingCommandType.SET_GOAL_AND_PATH);
